src/requirements.txt
-r test_requirements.txt

Django==2.1
djangorestframework==3.8.2
django-dotenv==1.4.2
psycopg2==2.7.5
celery==4.1.1
redis==2.10.5
